comb-to-cggi: fix secret cast lowerings to handle multi-dimensional tensors
Testcase describes how this appears in the tosa-to-boolean-tfhe pipeline.

I just have to decide whether to resolve the result shapes by allocating the memref of the correct shape in the RTLIL Importer instead of expand_shape op. It doesn't matter for the opernad shapes and collapse_shape because --fold-memref-alias-ops will get rid of the load/stores from the flat tensor. But it can't "remove" the flat allocated result tensor.

PiperOrigin-RevId: 655160200